# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/acl/packages-activate/": {
            "post": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Acl"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AclPackagesActivateController",
                "parameters": [
                    {
                        "description": "Add Role",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AclPackageActivate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AclResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/acl/packages-invoice/": {
            "post": {
                "description": "Description of the api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Acl"
                ],
                "summary": "Role Create Api Router",
                "operationId": "AclPackagesInvoiceController",
                "parameters": [
                    {
                        "description": "Add Role",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AclPackageInvoice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Roles",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/models.AclPackageInvoiceSuccess"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/libs.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "libs.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.AclPackageActivate": {
            "type": "object",
            "properties": {
                "custom": {
                    "type": "string"
                },
                "invoice": {
                    "type": "string"
                },
                "ipn_track_id": {
                    "type": "string"
                },
                "item_name": {
                    "type": "string"
                },
                "mc_amount_1": {
                    "type": "string"
                },
                "mc_amount_3": {
                    "type": "string"
                },
                "payer_email": {
                    "type": "string"
                },
                "payer_status": {
                    "type": "string"
                },
                "period_1": {
                    "type": "string"
                },
                "period_3": {
                    "type": "string"
                },
                "recur_times": {
                    "type": "string"
                },
                "recurring": {
                    "type": "string"
                },
                "subscr_date": {
                    "type": "string"
                },
                "subscr_id": {
                    "type": "string"
                },
                "txn_type": {
                    "type": "string"
                }
            }
        },
        "models.AclPackageInvoice": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "models.AclPackageInvoiceSuccess": {
            "type": "object",
            "properties": {
                "invoice": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}